vibe4trading-backend/pyproject.toml
[project]
name = "vibe4trading-backend"
version = "0.1.0"
description = "Vibe4Trading backend (internal pipeline modular monolith)."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "alembic>=1.13.2",
  "celery[redis]>=5.4.0",
  "defusedxml>=0.7.1",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "nautilus-trader>=1.223.0",
  "psycopg[binary]>=3.2.6",
  "pydantic>=2.7.4",
  "pydantic-settings>=2.4.0",
  "pystache>=0.6.5",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy>=2.0.31",
  "structlog>=24.4.0",
  "tenacity>=8.5.0",
  "uvicorn[standard]>=0.30.6",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/v4t"]

[dependency-groups]
dev = [
  "pyright>=1.1.389",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008", "E741"]

[tool.ruff.lint.isort]
known-first-party = ["v4t"]
